+++ src/main/java/com/orderfleet/webapp/web/rest/dto/LocationData.java
@@ -14,7 +14,7 @@ public class LocationData {
 
 		private String key;
 	
-		private Double battery_percentage;
+		private Integer battery_percentage;
 		
 		public String getCompanyName() {
 			return companyName;
@@ -56,18 +56,16 @@ public class LocationData {
 			this.key = key;
 		}
 
-		
-		
 
-		public Double getBattery_percentage() {
-			return battery_percentage;
-		}
+	public Integer getBattery_percentage() {
+		return battery_percentage;
+	}
 
-		public void setBattery_percentage(Double battery_percentage) {
-			this.battery_percentage = battery_percentage;
-		}
+	public void setBattery_percentage(Integer battery_percentage) {
+		this.battery_percentage = battery_percentage;
+	}
 
-		@Override
+	@Override
 		public String toString() {
 			return "LocationData [companyName=" + companyName + ", date=" + date + ", latitude=" + latitude
 					+ ", longitude=" + longitude + ", key=" + key + ", batteryPercentage=" + battery_percentage + "]";
